Adding check to install script to make sure it's not already installed

+++ install.sh
@@ -1,6 +1,12 @@
 #!/bin/zsh
 
-mv phpunit.sh ~
-chmod +x ~/phpunit.sh
-echo "alias phpunit='~/phpunit.sh'" >> ~/.zshrc
-source ~/.zshrc
+FILE=~/phpunit.sh
+
+if [ -f "$FILE" ]; then
+  echo "Looks like you've already ran this install script!"
+else
+  mv phpunit.sh ~
+  chmod +x ~/phpunit.sh
+  echo "alias phpunit='~/phpunit.sh'" >> ~/.zshrc
+  source ~/.zshrc
+fi
